Refactor interpolation/calibration to use double precision
- Changed all forward curve and interpolation data structures to use std::pair<double, double> for improved precision.
- Updated Interpol class constructor and setParams to accept double-based forward curves.
- Modified interpolate1D and makeInterpolator utilities for double keys.
- Extended HaganCalibrationResult with optimal_params and calibration_results fields.
- Refactored Interpol::calibrate_hagan: expiry is now double, market_vols is a vector of double pairs.
- Added new calibration.cpp with coordinate descent SABR calibration and detailed diagnostics.
- Updated project file to use calibration.cpp instead of old calibration_hagan.cpp.
- Improves numerical accuracy and calibration reporting.

+++ INTERPOL/interpol.cpp
@@ -6,22 +6,22 @@
 #include <functional>
 
 // Flat curve for default setting
-std::vector<std::pair<int, double>> flat_curve = []() {
-    std::vector<std::pair<int, double>> curve;
+std::vector<std::pair<double, double>> flat_curve = []() {
+    std::vector<std::pair<double, double>> curve;
     for (int i = 1; i <= 100; ++i) {
-        curve.push_back({ i, 100.0 });
+        curve.push_back({ static_cast<double>(i), 100.0 });
     }
     return curve;
 }();
 
 // Constructor
-Interpol::Interpol(double alpha_, double beta_, double gamma_, std::vector<std::pair<int, double>> fwd_curve_)
+Interpol::Interpol(double alpha_, double beta_, double gamma_, std::vector<std::pair<double, double>> fwd_curve_)
     : alpha(alpha_), beta(beta_), gamma(gamma_), fwd_curve(makeInterpolator(fwd_curve_))
 {
 }
 
 // Set parameters
-void Interpol::setParams(double alpha_, double beta_, double gamma_, double rho_, std::vector<std::pair<int, double>> fwd_curve_)
+void Interpol::setParams(double alpha_, double beta_, double gamma_, double rho_, std::vector<std::pair<double, double>> fwd_curve_)
 {
     alpha     = alpha_;
     beta      = beta_;
